+++ src/main/java/com/sedin/qna/account/controller/AccountController.java
@@ -1,16 +1,15 @@
 package com.sedin.qna.account.controller;
 
-import com.sedin.qna.account.model.Account;
 import com.sedin.qna.account.model.AccountDto;
+import com.sedin.qna.account.model.CurrentAccount;
 import com.sedin.qna.account.service.AccountService;
-import com.sedin.qna.common.LoginRequired;
-import com.sedin.qna.network.ApiResponseDto;
+import com.sedin.qna.common.response.ApiResponseDto;
+import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestAttribute;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseStatus;
@@ -19,14 +18,11 @@ import org.springframework.web.bind.annotation.RestController;
 import javax.validation.Valid;
 
 @RestController
-@RequestMapping("api/accounts")
+@RequiredArgsConstructor
+@RequestMapping("/api/accounts")
 public class AccountController {
 
     private final AccountService accountService;
-
-    public AccountController(AccountService accountService) {
-        this.accountService = accountService;
-    }
 
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
@@ -34,21 +30,18 @@ public class AccountController {
         return ApiResponseDto.OK(new AccountDto.ResponseOne(accountService.signUp(create)));
     }
 
-    @PatchMapping("/{id}")
+    @PatchMapping
     @ResponseStatus(HttpStatus.OK)
-    @LoginRequired
-    public ApiResponseDto<AccountDto.ResponseOne> update(@RequestAttribute Account account,
-                                                         @PathVariable Long id,
+    public ApiResponseDto<AccountDto.ResponseOne> update(@CurrentAccount String email,
                                                          @RequestBody @Valid AccountDto.Update update) {
-        return ApiResponseDto.OK(new AccountDto.ResponseOne(accountService.update(account, id, update)));
+        return ApiResponseDto.OK(new AccountDto.ResponseOne(accountService.update(email, update)));
     }
 
-    @DeleteMapping("/{id}")
+    @DeleteMapping
     @ResponseStatus(HttpStatus.OK)
-    @LoginRequired
-    public ApiResponseDto<String> delete(@RequestAttribute Account account,
-                                         @PathVariable Long id) {
-        accountService.delete(account, id);
+    public ApiResponseDto<String> delete(@AuthenticationPrincipal String email) {
+        accountService.delete(email);
         return ApiResponseDto.DEFAULT_OK;
     }
+
 }
