+++ src/main/java/com/sedin/qna/account/repository/AccountRepository.java
@@ -7,9 +7,8 @@ import java.util.Optional;
 
 public interface AccountRepository extends JpaRepository<Account, Long> {
 
-    Optional<Account> findByLoginId(String loginId);
-
-    boolean existsByLoginId(String loginId);
-
     boolean existsByEmail(String email);
+
+    Optional<Account> findByEmail(String email);
+
 }
