+++ src/main/java/com/sedin/qna/account/service/AccountService.java
@@ -1,13 +1,12 @@
 package com.sedin.qna.account.service;
 
-import com.sedin.qna.account.model.Account;
 import com.sedin.qna.account.model.AccountDto;
 
 public interface AccountService {
 
     AccountDto.Response signUp(AccountDto.Create create);
 
-    AccountDto.Response update(Account account, Long id, AccountDto.Update update);
+    AccountDto.Response update(String email, AccountDto.Update update);
 
-    void delete(Account account, Long id);
+    void delete(String email);
 }
